clean up for the version with generalized power cones
