+++ src/Clarabel.jl
@@ -63,8 +63,6 @@ module Clarabel
     include("./cones/coneops_expcone.jl")
     include("./cones/coneops_powcone.jl")
     include("./cones/coneops_genpowcone.jl")        #Generalized power cone 
-    include("./cones/coneops_powmeancone.jl")       #Power mean cone
-    include("./cones/coneops_entropycone.jl")       #Relative entropy cone
     include("./cones/coneops_compositecone.jl")
     include("./cones/coneops_exppow_common.jl")
     include("./cones/coneops_symmetric_common.jl")
