+++ src/MOI_wrapper/MOI_wrapper.jl
@@ -18,9 +18,7 @@ const OptimizerSupportedMOICones{T} = Union{
     MOI.ScaledPositiveSemidefiniteConeTriangle,
     MOI.ExponentialCone,
     MOI.PowerCone{T},
-    Clarabel.GenPowerConeT,
-    Clarabel.PowerMeanConeT,
-    Clarabel.EntropyConeT,
+    Clarabel.GenPowerConeT
 } where {T}
 
 Base.copy(cone::OptimizerSupportedMOICones) = cone 
@@ -635,17 +633,9 @@ function push_constraint_set!(
         return nothing
     end
 
-    # handle self-defined GenPowerCone & Relative Entropy cone & Power mean cone
+    # handle self-defined GenPowerCone
     if isa(s,Clarabel.GenPowerConeT)
         push!(cone_spec, Clarabel.GenPowerConeT(s.α,s.dim1,s.dim2))
-        return nothing
-    end
-    if isa(s,Clarabel.PowerMeanConeT)
-        push!(cone_spec, Clarabel.PowerMeanConeT(s.α,s.d))
-        return nothing
-    end
-    if isa(s,Clarabel.EntropyConeT)
-        push!(cone_spec, Clarabel.EntropyConeT(s.dim))
         return nothing
     end
