+++ src/cones/compositecone_type.jl
@@ -48,12 +48,6 @@ struct CompositeCone{T} <: AbstractCone{T}
             elseif typeT == GenPowerConeT
                 cones[i] = ConeDict[typeof(cone_specs[i])]{T}(T.(cone_specs[i].α),cone_specs[i].dim1,cone_specs[i].dim2)
                 _is_symmetric = false
-            elseif typeT == PowerMeanConeT
-                cones[i] = ConeDict[typeof(cone_specs[i])]{T}(T.(cone_specs[i].α),cone_specs[i].d)
-                _is_symmetric = false
-            elseif typeT == EntropyConeT
-                cones[i] = ConeDict[typeof(cone_specs[i])]{T}(cone_specs[i].dim)
-                _is_symmetric = false
             else
                 cones[i] = ConeDict[typeT]{T}(cone_specs[i].dim)
             end
